requirements.txt
Pillow>=9.0.0
pydub>=0.25.1
transformers>=4.0
torch>=1.0
torchaudio>=0.10.0
soundfile>=0.12.0
hydra-core>=1.3.2
omegaconf>=2.3.0
#git+https://github.com/nari-labs/dia.git
pytest>=7.0
black>=22.0
flake8>=4.0
python-dotenv>=1.0.0
requests>=2.31.0